+++ src/orchestrator/deploy/local.go
@@ -17,19 +17,18 @@ type localImpl struct {
 	PathFile               string `json:"pathFile"`
 	NbInstanceToLaunch     int    `json:"nbInstanceToLaunch"`
 	CleanPreviousInstances bool   `json:"cleanPreviousInstances"`
-	LocalIP                string `json:"localIp"`
 	Port                   int    `json:"port"`
 
-	consulAddress string
-	cmdFabric     func(name string, arg ...string) *exec.Cmd
+	mqttAddress string
+	cmdFabric   func(name string, arg ...string) *exec.Cmd
 }
 
-func newLocalFromJSON(serialized json.RawMessage, consulClient tools.Consul) (deployer, error) {
+func newLocalFromJSON(serialized json.RawMessage, mqttClient tools.Mqtt) (deployer, error) {
 	d := &localImpl{}
 	if err := json.Unmarshal(serialized, d); err != nil {
 		return nil, err
 	}
-	d.consulAddress = consulClient.GetAddress()
+	d.mqttAddress = mqttClient.GetAddress()
 	d.cmdFabric = exec.Command
 	return d, nil
 }
@@ -49,10 +48,7 @@ func (local *localImpl) Deploy() error {
 
 	for i := 0; i < local.NbInstanceToLaunch; i++ {
 		go func() {
-			args := []string{"-consul", local.consulAddress}
-			if local.LocalIP != "" {
-				args = append(args, "-local-ip", local.LocalIP)
-			}
+			args := []string{"-mqtt", local.mqttAddress}
 			if local.Port != 0 {
 				args = append(args, "-port", strconv.Itoa(local.Port))
 			}