+++ src/orchestrator/prometheus/prometheus_test.go
@@ -3,26 +3,12 @@ package prometheus
 import (
 	"context"
 	"errors"
-	"lorhammer/src/tools"
 	"testing"
 	"time"
 
 	api "github.com/prometheus/client_golang/api/prometheus/v1"
 	"github.com/prometheus/common/model"
 )
-
-type fakeConsul struct {
-	serviceFirstURL   string
-	serviceFirstError error
-}
-
-func (fakeConsul) GetAddress() string                                      { return "" }
-func (fakeConsul) Register(ip string, hostname string, httpPort int) error { return nil }
-func (f fakeConsul) ServiceFirst(name string, prefix string) (string, error) {
-	return f.serviceFirstURL, f.serviceFirstError
-}
-func (fakeConsul) DeRegister(string) error                     { return nil }
-func (fakeConsul) AllServices() ([]tools.ConsulService, error) { return nil, nil }
 
 type fakePrometheusAPI struct {
 	results      []float64
@@ -45,7 +31,7 @@ func (f fakePrometheusAPI) LabelValues(ctx context.Context, label string) (model
 }
 
 func TestNewApiClient(t *testing.T) {
-	c, err := NewAPIClientFromConsul(fakeConsul{})
+	c, err := NewAPIClient("")
 	if err != nil {
 		t.Fatal("Valid prometheus config should not return error")
 	}
@@ -54,18 +40,8 @@ func TestNewApiClient(t *testing.T) {
 	}
 }
 
-func TestNewApiClientErrorConsul(t *testing.T) {
-	c, err := NewAPIClientFromConsul(fakeConsul{serviceFirstError: errors.New("error")})
-	if err == nil {
-		t.Fatal("No prometheus found in consul should return error")
-	}
-	if c != nil {
-		t.Fatal("No prometheus found in consul should not return api client")
-	}
-}
-
 func TestNewApiClientErrorBadConsulUrl(t *testing.T) {
-	c, err := NewAPIClientFromConsul(fakeConsul{serviceFirstURL: ":"})
+	c, err := NewAPIClient(":")
 	if err == nil {
 		t.Fatal("Bad url for prometheus in consul should return error")
 	}
@@ -75,7 +51,7 @@ func TestNewApiClientErrorBadConsulUrl(t *testing.T) {
 }
 
 func TestApiClientImpl_ExecQuery(t *testing.T) {
-	c, _ := NewAPIClientFromConsul(fakeConsul{})
+	c, _ := NewAPIClient("")
 	c.(*apiClientImpl).queryAPI = fakePrometheusAPI{results: []float64{255.0}}
 	res, err := c.ExecQuery("")
 	if err != nil {
@@ -88,7 +64,7 @@ func TestApiClientImpl_ExecQuery(t *testing.T) {
 }
 
 func TestApiClientImpl_ExecQueryZero(t *testing.T) {
-	c, _ := NewAPIClientFromConsul(fakeConsul{})
+	c, _ := NewAPIClient("")
 	c.(*apiClientImpl).queryAPI = fakePrometheusAPI{results: []float64{}}
 	res, err := c.ExecQuery("")
 	if err != nil {
@@ -100,7 +76,7 @@ func TestApiClientImpl_ExecQueryZero(t *testing.T) {
 }
 
 func TestApiClientImpl_ExecQueryError(t *testing.T) {
-	c, _ := NewAPIClientFromConsul(fakeConsul{})
+	c, _ := NewAPIClient("")
 	c.(*apiClientImpl).queryAPI = fakePrometheusAPI{results: []float64{}, resultsError: errors.New("error")}
 	_, err := c.ExecQuery("")
 	if err == nil {