+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.6.0 - UNRELEASED
 
+* **CONSUL** remove consul and pass mqtt url directly, permit to deploy more easily and to launch multiple orchestrator [issues/66](https://gitlab.com/itk.fr/lorhammer/issues/66)
 * **LORHAMMER** allow to set port used to be scrapped by prometheus [issues/64](https://gitlab.com/itk.fr/lorhammer/issues/64)
 * **ORCHESTRATOR** exit(1) if an error occur during scenario (previously exit > 0 only when checker fail) [issues/67](https://gitlab.com/itk.fr/lorhammer/issues/67) + repair bug on ramp type test
 * **CHECKER** fix bug mqtt checker not working [issues/62](https://gitlab.com/itk.fr/lorhammer/issues/62)